do not mask and shift TxClampConfig data
The datasheet states in 15.2.2 that the `TxClampConfig` register should
be modified to optimize the PA clamping threshold.
For that the user shall read the register, OR it with 0x1E and write it
back.

With the current implementation this is not possible since it applies a
mask (0x0F) on the byte, thinking only bits 4:1 would be relevant.
This is not correct, for the optimized clamping bits 4:1 should be set
but the other bits continue to be relevant (as long as the pseudo code
is correct).
The default of this register is 0xC8 which is the next sign that the
other bits are relevant.

Current implementation:
    1. Read: (0xC8 & 0x0F) >> 1 --> result: 0x04
    2. Apply workaround: threshold |= 0x1E -> threshold: 0x1E
    3. Write: (0x1E & 0x0F) << 1 --> result: 0x1C
       (using 0x0F would result in 0x1E instead of 0x1C)

Following the pseudo code the correct value would be:
    0xC8 | 0x1E = 0xDE

So this commit removes all masking and shifting and gives the user the
full control.